newsletter-agent/requirements.txt
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pyyaml>=6.0.1
typer>=0.12.3
rich>=13.7.1
httpx>=0.27.0
tenacity>=8.2.3
anthropic>=0.25.6 ; python_version >= "3.8"
openai>=1.35.9 ; python_version >= "3.8"
python-slugify>=8.0.4
rapidfuzz>=3.6.2
youtube-transcript-api>=0.6.2
